backend/requirements.txt
asgiref==3.7.2
black==23.11.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.7
django-environ==0.11.2
django-filter==23.4
django-leaflet==0.29.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-gis==1.0
drf-yasg==1.21.7
flake8==6.1.0
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
mypy-extensions==1.0.0
openapi-codec==1.3.2
packaging==23.2
pathspec==0.11.2
platformdirs==4.0.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.1.0
pyscopg2==66.0.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
simplejson==3.19.2
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.1.0
